crostini: Make Settings.GetSharedFolders() more reliable
This CL adjust the logic of Settings.GetSharedFolders() so that it is
more reliable when run immediately after settings is opened. When
opened, the shared folders page does an async request to load the actual
list of shared folders. The current code can race with this and end up
returning an error as a result.

Before the async request completes, the page displays both the nodes
matching emptySharedFoldersMsg and sharedFoldersList, and after it
completes only one of these will be present. The new code waits for the
latter condition before trying to get the list of shared folders.

BUG=chromium:1202633
TEST=run with test in http://crrev.com/i/3884691

Change-Id: I02acb8df83c57075d1917140072b261f0b071fee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2948486
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>