tast: Add simpler way to initialize helper
A typical test that uses servo will use the helper with the same
boilerplate. Added a NewHelperS() constructor to make that simpler.

A test that powers off the DUT should really power it back up at the end
of the test, and CheckPowerAndClose() will do that. If the DUT is up, it
will do nothing so it is safe to call via defer.

BUG=b:190532175
TEST=tast run --var=servo=localhost:9999 \
  --var=firmware.hibernate_time=1m $DUT \
  firmware.DeepSleep

Change-Id: I786d9167ac5f7930b4d4bd254b4b3a9c0ef18dca