tast: Improvements to firmware.DeepSleep
Switch from servo proxy to the firmware.Helper provided by pre.Helper.

Boot the DUT if it is off when the test fails.

Verify that the EC is actual hibernating, and fail the test if it is not.

Instead of waking the DUT by attaching the charger, use the power key.
Otherwise the battery has a chance to charge a little bit before it is
measured. Instead we will get the most pessimistic value for the battery
life.

BUG=b:190532175
TEST=tast run --var=servo=localhost:9999 \
  --var=firmware.hibernate_time=20m $DUT \
  firmware.DeepSleep

Change-Id: I36967081647b6e595896c08cfc13b4721ee0581c