tast: Improvements to firmware.DeepSleep
Switch from servo proxy to the firmware.Helper provided by
pre.NormalMode.

Verify that the EC is actual hibernating, and fail the test if it is not.

Instead of waking the DUT by attaching the charger, use the power key.
Otherwise the battery has a chance to charge a little bit before it is
measured. Instead we will get the most pessimistic value for the battery
life.

If the test runs, and the starting battery level is the same as the
ending battery level, then it has infinite battery life, or rather, we
don't have enough information to verify. If the battery life was exactly
100 days, then the minimum time to lose 1mAh would be
(100d / battery capacity). If the test has run for less than that, then
fail the test with a message to increase the test time.

Unless the battery has a capicity of 600mAh or less, the default 4h
sleep time won't trigger this condition, but manual overrides of
var firmware.hibernate_time will.

Instead of a long power key press, use hold_pwr_button_poweroff from the
fw-testing-configs, when shutting down the DUT in firmware.DeepSleep

BUG=b:190532175
TEST=tast run --var=servo=localhost:9999 \
  --var=firmware.hibernate_time=10s $DUT \
  firmware.DeepSleep

Change-Id: I36967081647b6e595896c08cfc13b4721ee0581c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2951031
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>