[Lacros] Update lacros.AppLauncherLaunch tast test.
This CL updates the lacros.AppLauncherLaunch tast test to work with
keeping lacros alive in the background with AppService.

When app service started it will start the lacros in the background and
create the user directory, and the existing code will remove the
directory after this, and starting Lacros Browser after this doesn't
create new user directory, which breaks the test. This CL updates the
data clear before creating new Lacros session.

BUG=chromium:1144877
TEST=Manually tested lacros.AppLauncherLaunch with the keepalive CL
(https://crrev.com/c/2936355) patched.

Change-Id: I1d1f814c9d6d0a9fa11c41ef4db109b98b993133