video.DrawOnCanvas: Use requestVideoFrameCallback.
I started noticing some flakiness in the color distances calculated by
this test. This is reminiscent of https://crbug.com/1162412 where it was
suggested that the right thing to do was to use
video.requestVideoFrameCallback(). This CL makes use of that in order to
know when we can draw the video on the canvas in an attempt to eliminate
flakiness.

BUG=b:159158467
TEST=tast run ${IP} video.DrawOnCanvas.h264_360p_hw 50 times on ezkinil

Change-Id: I8ffcd530ad61a0044ca6b9a230ba594912ccabeb