tast-tests: Fail DeepSleep if time too short
If the test runs, and the starting battery level is the same as the
ending battery level, then it has infinite battery life, or rather, we
don't have enough information to verify. If the battery life was exactly
100 days, then the minimum time to lose 1mAh would be
(100d / battery capacity). If the test has run for less than that, then
fail the test with a message to increase the test time.

Unless the battery has a capicity of 600mAh or less, the default 4h
sleep time won't trigger this condition, but manual overrides of
var firmware.hibernate_time will.

BUG=b:190532175
TEST=tast run --var=servo=localhost:9999 \
  --var=firmware.hibernate_time=10s $DUT \
  firmware.DeepSleep

Change-Id: I1a0a58544fa3b4cb1b507b9deb4c8ea1c822cfa2