tast-tests: After the test, use dut.WaitConnect
Instead of failing if the DUT is down, wait for it to finish booting.
Many tests, particularly in failure cases leave the DUT in the process
of booting, and it might take some retries before it can be connected.

BUG=None
TEST=tast run --var=servo=localhost:9999 \
  --var=firmware.hibernate_time=10s $DUT \
   firmware.BootTime* firmware.DeepSleep
For my DUT, with these flags, both tests fail. Without this change,
the second test doesn't run at all.

Change-Id: I4662d1df1c54bd72646cbbe5bf78321cfe6442fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2966924
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>