tast: multi_networking_connectivity: Rework the test for ARCVM
Prior to this patch the test does:
1. Ping gateway address of each physical interface from a
   specific ARC interface (using ping -I option).
2. Ping ARC interface from the host and verify that the ping
   is coming from the right interface (using tcpdump).

ARCVM currently does not support commands running from shell
with root permission, without doing 'adb root' (which requires
the build to be a test build).

This patch removes "-I" from (1) and "tcpdump" from (2).
This patch allows the test to be run for both ARC and ARCVM.

BUG=b:168011514
TEST=tast run <DUT> arc.MultiNetworkingConnectivity
TEST=tast run <DUT> arc.MultiNetworkingConnectivity.vm

Change-Id: Ie4538049a28ea2993fdd8aa9d1d86a8fee1e8801