tast-tests: Capture servod log at end of test
SSH to the servo host, and download the servod DEBUG log and store it
in the outDir of the test. This is the most important log that is
present in Tauto, but missing in tast.

BUG=b:185998785
TEST=firmware.Pre* in 3 modes:
local - port
remote - with ssh tunnel for servo port
remote - with real servo port, tunneled ssh port

Change-Id: I4912d8c5274cc623a449c3dc94b666b74792491b