feedback.SysInfoPII: don't depend on network.
The flakiness in the linked bug was caused by our network dependency.
Specifically, when the network was down, the page title would end up
being "www.google.com", which is incidentally logged by the captive
portal detection code.

In a future CL, we'll add logic to ensure that arbitrary hostnames are
not logged.

We no longer treat file:// URLs specially (as of
https://crrev.com/c/2770527) so it's now fine to test using a local
file, avoiding the inherent flakiness that comes from relying on a
stable internet connection.

BUG=chromium:1218182
TEST=tast run $DUT feedback.SysInfoPII

Change-Id: I6b323306c604e3c054a113c6587d47025a2cfb3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2964319
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>