crostini: Retry unsharing of folders upon failure
The shared folders settings page shows a modal dialog if it was unable
to unshare a folder due to a file being in use by the guest. This causes
flakiness in pita.FolderSharing, where we read the contents from a file
and immediately try to unshare.

The current behaviour simply leaves the dialog there, although it tries
to click OK to dismiss the dialog. Clicking OK is not a good choice as
while the entry will be removed from the list, the file will remain
shared until either the VM is restarted or the user is logged out of
Chrome. This CL changes the behaviour to instead click the 'Try again'
button a couple of times, and then return an error if the dialog does
not go away.

BUG=chromium:1202633
TEST=tast run $DEVICE_IP crostini.Share*.buster_stable
TEST=Run pita.FolderSharing tast test

Change-Id: I6f1a9c57e9c04959cf9a45dea85c79c5b085298e