tast-tests: Ignore gRPC cancel errors in close
When the DUT shuts down, the ssh connection naturally dies, and the RPC
connection fails as well. When closing it, gRPC returns a cancelled
error, but that is not really an error in the close case, so ignore it.

Unfortunately we need to unwrap the error ourselves, as there isn't a
good way to get the errors package and status.FromError() to play
together.

BUG=b:185271203
TEST=tast run firmware.ServoGBBFlags

Change-Id: I4006a13fd1727656c64cf36d01033076a8d78e98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2967124
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>