feedback: Add SysInfoPII subtest for hostnames.
As part of the linked bug, we noticed that we're not testing to ensure
hostnames are not logged in syslogs.

Add a test that verifies this condition, which uses a third-party site
because there are limited cases where logging first-party hostnames is
acceptable (for instance, shill's captive portal detection logs
"www.google.com" in some cases, but this is not a response to any user
actions; rather, it's a fixed constant).

BUG=chromium:1218182
TEST=tast run $DUT feedback.SysInfoPII*

Change-Id: I6938a9dd41168727b16bccdddc57ec50f87795a2