ippusbbridge: Add a function to check socket connectivity
The ippusbbridge.WaitForSocket function lets a test wait for the
ippusb_bridge socket to appear in the filesystem, but that doesn't
guarantee that traffic will reach the device on the other side.  Add a
new CheckDevice function that sends an HTTP request through the socket
to ensure that a particular URL is reachable.  This should reduce test
flakes caused by slow USB enumeration, and will also provide a clear
error message instead of a timeout when IPP-USB communication isn't
working.

BUG=b:188590605, b:188591629
TEST=20 cycles of tast run scanapp.* documentscanapi.Scan

Change-Id: I6c5dcd8abdbc062cd3046a085d0702e1e27a3d18