tast-tests: Update scrollForward and scrollBackward method
Some of the tast.arcappcompat tests are failing in Clamshell: ScrollForward test case due to error mentioned in the bug, b:191289287. Observed that the error occurs even though the scrollForward method performs scroll action on the app page.

Found the root cause of the failure, callSimple method interprets the output from scrollforward as success or failure case. But as per scrollForward and scrollBackward method in UIAutomator server (AutomatorService.java#L572 provided in b:191289287), will return output as true if it is scrolled or false if the page is not scrollable any more.  Currently tast UI Object doesn't have a method to interpret the outputs of scrollForward and scrollBackward methods in the UIAutomator server. Implemented a new method which will interpret the outputs of scrollForward and scrollBackward methods correctly and also will fix the flaky failures.

BUG=b:191289287
TEST=tast run <eve> arcappcompat.IHeartRadio

Change-Id: Ife1c7ff1712bd1f39017747996d5317d89cb9458