tast-tests: Ensure arc.PIPRoundedCornersUnderlay* test what they should
The test is supposed to check that if the ARC++ PIP video is promoted to
overlay, then it uses the underlay overlay strategy. Although there may
be legitimate reasons for the ARC++ PIP video not to be promoted to
overlay, sometimes it just takes a moment to be promoted to overlay, and
the test should wait to see what overlay strategy is used then. The
present CL addresses that by waiting for the app to idle.

BUG=chromium:1107460
TEST=arc.PIPRoundedCornersUnderlay and arc.PIPRoundedCornersUnderlay.vm

Change-Id: I858de239e1b078bc23277cf4ac0804eb473051fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2969725
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Avery Musbach <amusbach@chromium.org>