crash: Fix slowdown introduced by logging.
For reasons that aren't quite clear to me, ioutil.ReadAll always blocks
the full 30 seconds that the subcommand is scheduled to run -- even
though Wait returns much more quickly, and the subcommand is killed
quickly as well. Work around this by using io.Read() rather than
ioutil.ReadAll, and accepting that we may not get the full log.

BUG=b:191337331
TEST=time ../platform/tast/fast_build.sh -t chromiumos/tast/local/crash -- -count=1

Change-Id: I95a07b2a0344f697265af5c8b9055efb14bb38f2