tast-test: Bring system-proxy notification to UI foreground
If a system service or another ARC app is performing a network request
through system-proxy before the test app is installed, the system-proxy
notification is shown.

After the app is installed, it will be shown in full screen, covering
the system-proxy notification. If the notification is not visible, the
uig utility can't find it, causing the network.SystemProxyForArc test
to fail.

This CL ensures the notification is always on top by clicking on the
tray bar.

BUG=b:190206838
TEST=tast -verbose run ${DUT} network.SystemProxyForArc + manually
  forcing the notification before the ARC test app is installed by
  doing a curl connection to system-proxy

Change-Id: If711799651f551706b606b99f78848f2a9482356