shill/fixture.go: extend timeout to wait for services
This change modifies the timeout of the resetShill function to be the
context's deadline. This both extends the timeout and makes future
changes easier since only the fixture timeouts have to be modified.

resetShill was reliably timing out on lower end devices (sentry) taking
about ~17 seconds to complete. The new 30 second timeout should be
enough.

BUG=chromium:1221899
TEST=tast run sentry-dut network.DiagFailLANConnectivity (pass)

Change-Id: Ie5a250d8f994a2acc46d96754c15606762b30be3