smaps pss_total metrics
The pss and pss_swap metrics are unstable, possible because it's a bit
random which set of processes get reclaimed. Make a new metric,
pss_total which is the sum of pss and pss_swap, that will be unaffected
by which pages are reclaimed.

BUG=b:178987626
TEST=tast run dut multivm.Login.arc

Change-Id: I9c043a57cf08de9edb7b33ec293e79ad60e5a08b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2987727
Commit-Queue: Charles William Dick <cwd@google.com>
Tested-by: Charles William Dick <cwd@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>