tast-tests: Remove call to deprecated function
As part of the effort to migrate tests from the chrome/ui package to
the chrome/uiauto package, this CL removes a call to a function in the
chrome/ui package. This call was originally written out of concern for
the possibility that the PIP window resize could lag behind the resizing
drag. That would be no big disaster, and it probably will never happen.

BUG=b:190614781
TEST=tast run <DUT> arc.PIPEnergyAndPower.big*

Change-Id: I1a97a7c11c56c424d9cd22da151cb6c3520ad2d0