tast-tests: Revert recent change to activity.ResizeWindow
The present CL reverts a part of [1] which changed activity.ResizeWindow
to drag the corner instead of the resize handle. It forestalled flake in
the added tests, but it also eliminated test coverage of the resize
handle, as activity.ResizeWindow is also used by arc.PIP and arc.PIP.vm.
With [2], arc.PIPEnergyAndPower.big* will not use activity.ResizeWindow.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2946092
[2] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2976673

BUG=b:168011269
TEST=tast run <DUT> arc.PIP and tast run <DUT> arc.PIP.vm

Change-Id: Idab143e4d7d31f9d7f760362276336cd7e092cb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2976111
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>