tast-tests: syzcorpus: kill process after RunRepro
Some repros continue to run in the background even after the timeout.
Issue a pkill to kill any such processes that might be running. Killing
any repros running in the background also resolves spurious reboots and
DUT slowdown while testing; the DUT does not have to Reboot() at the
end of the test.

Also, clear the dmesg log before starting to test.

BUG=b:189825001
TEST=tast -verbose run localhost:2200 syzcorpus.Kvmtest

Change-Id: I681ec8242c83a100f21c75151c21e9cfcf819d9f