arc.Print: Switch to arcBooted fixture
This test isn't passing any custom flags to Chrome or Arc, so it doesn't
need to individually set up a session.  It was switched a while back to
avoid flakes caused by other informational tests, but this should no
longer be a concern now that arc.Print is in the CQ.  Switching to the
shared fixture also gets us things like automatically disabling the
ArcResizeLock feature when needed.

BUG=b:191701248
TEST=arc.Print, arc.Print.unstable, arc.Print.vm on various boards

Change-Id: Ie1e72f9ea5931da02c18e70a1c138a29829e8106