network: VPNConnect: Add sleep before StartL2TPIPsecServer() returns
We observed sometimes the L2TP/IPsec test failed with message
NO_PROPOSAL_CHOSEN. The reason should be that the client side sends the
connection request before the server side loads the corresponding
configurations. Since there is no direct way to know when the
configurations are read by the server, this patch adds a sleep to
mitigate this issue.

BUG=b:183482749
TEST=tast run ${DUT} network.VPNConnect.l2tp_ipsec_psk

Change-Id: I3ee5146de83e04e0d4de699ccef301107730ea35