upstart: support multiple-instance jobs
This CL expands upstart lib to support multiple-instance jobs. It adds
`args ...Arg` to several functions, enable multiple-instance jobs to
start, stop and check status of a specific instance.

This change is transparent to customers of upstart lib, just to support
future usage where a multiple-instance job wants to use upstart in tast.

BUG=b:191731226
TEST=./fast_build.sh -T; tast run platform.MLServiceBootstrap with
TEST=other WIP CL that makes ml-service multiple-instance

Change-Id: I61117a8bead99c66c0cf64c9ec8a8a7d3a8ab980