network: DNSProxyOverVPN: Ensures DNS proxy over VPN works
DNS proxy over VPN test check if DNS functionality are
working properly over VPN when DNS proxies are running.
The test now checks 3 scenarios:
1. DNS queries should be successful on a VPN.
2. DNS queries should fail on VPN if DNS on the VPN
   server is blocked (confirms that the queries got
   routed through the VPN server).
3. DNS queries on always on mode should be successful
   with VPN secure DNS blocked (confirms that DoH is
   disabled on VPN for default and ARC proxy).

The tests are done on system, user, and chrome as its
client.

The tests are also done with different DoH modes (Off,
Automatic, and Always On).

BUG=b:180551432
TEST=tast run ${DUT} network.DNSProxyOverVPN.*

Change-Id: I45c6581f9e024984882e08a7cc8671146ce98e42