network: Run command in a specific netns in RunChroot()
Currently, RunChroot() run the command by `minijail0 -e` which creates
a new network namespace, while in most cases, we want to run the command
in the netns created with this chroot. This patch changes this behavior
to the latter one by running it with `ip netns exec`.

BUG=b:191830846
TEST=tast run $DUT network.VPNConnect.*

Change-Id: I8b7b2f50de9be6e173f173d3be320c5a0523c42f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2982051
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>