(Nearby Share) Reconnect to Snippet server if it drops during login.
On some devices in our lab (e.g sarien), the DUT loses it's connection
to the Snippet Server running on the Android phone after a CrOS login.
This doesn't happen on other device (e.g coral, hatch, octopus).

To work around this issue we will check that the connection is still
active after logging in on ChromeOS and reconnect to the snippet server
if the connection was lost.

In order to debug this failure I also needed to add extra logging to our
fixtures.

BUG=b:191276960
TEST=group:nearby-share on Sarien/Pixel4a combo in lab

Change-Id: I785268994f56859b991f8ed6dc84d278730436fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2982314
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>