Fix clicking non-exist button in VK voice tests
Last CL reduces most of the flakiness in VK voice testing.
Besides the known VK-not-shown issue, another cause of failures is
that it fails to click "Loopback" nodes.

The reason is mostly because there is no guarantee that the nodes exist
before DoDefault actions.

Add a WaitUntilExists method to make sure the nodes exist.

BUG=b:188489307
TEST=tast run <DUT> inputs.VirtualKeyboardSpeech*

Change-Id: Ia0ea1e5e2d16a2350c67df1abcba9472cc4e60c2