tast-tests: Unflake spoken feedback test
The spoken feedback test has been failing recently [1] due to
changes in the way tabs are handled in ChromeVox. To unflake the
test, we open a web page, wait for ChromeVox to focus the root web
area, then begin peforming commands and expecting speech. This has
a few benefits:

1. It stabilizes the test
2. It follows the pattern of ChromeVox JavaScript tests
3. It exercises ChromeVox navigation logic

I ran the test 10 times in succession and observed no failures.

[1] https://tastboard.corp.google.com/test?contact=&testname=
a11y&dep=

BUG=chromium:1159107
TEST=for i in {1..10}; do tast run localhost:2224 a11y.SpokenFeedback.*; done

Change-Id: I7742962a88511021ce2674bb79390c31818988d3