tast-tests: Ensure that TTS test verifies event was dispatched
This CL is a follow-up to crrev/c/2935639, which added the initial TTS
test. This CL adds a check to validate the end event callback. It also
means that we can remove the poll when trying to read the spoken text
from a file.

BUG=b:168671331
TEST=arc.TextToSpeech

Change-Id: I6eb03a65c4478aa22083ce9e01414a2d8307a8a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2984280
Auto-Submit: Sara Kato <sarakato@chromium.org>
Tested-by: Sara Kato <sarakato@chromium.org>
Commit-Queue: Sara Kato <sarakato@chromium.org>
Reviewed-by: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>