tast-tests: Unflake ChromeVox hint test
The ChromeVox hint feature gives two different utterances, one
for laptop devices and another for tablet devices. The test fails
for tablet devices [1] because the test only expects the laptop
string to be spoken.

To stabilize this test, we add the ability for the speech monitor
to match regular expressions. This gives more flexibility, since
we no longer have to match strings exactly. This is good for
long-term stability, since strings are subject to change.

I verified that the test passes for both tablets and laptops by
using my Eve device to test in both form factors.

[1] https://tastboard.corp.google.com/test/a11y.ChromevoxHint

BUG=chromium:1159107
TEST=tast -verbose run localhost:2224 a11y.ChromevoxHint

Change-Id: I5e44127bec8b140cc152b5f3dfc12e052eb939d2