platform.CrosRuntimeProbe: skip tests when a category is waived
If we waive some categories, there will not exist such labels in the
host info labels.  Therefore, it should be passed in this case. This CL
makes tast tests skip the probe action and pass tests directly.

For tests with multiple categories, i.e. input_device and network, the
probe action will be executed as before.

BUG=b:192028575
TEST=run all tests on FLEEX

Change-Id: Id70e8b3a9f8ad41342bd4f7ef5a1dffc02914eb0