tast-tests: Port ui.ChromePIPEnergyAndPower.* to lacros
For simplicity, all 8 tests (originals and added lacros variations) will
use fixtures that log in to Chrome. The motivation for avoiding such
fixtures was that in the ChromeOS lab, the tests were failing when run
among other tests, but passing when run after a reboot. The root cause
was never identified, but I can hope things have changed so that this CL
will not reintroduce the problem. There is only one way to find out :-)

BUG=chromium:1093933
TEST=tast run <DUT> ui.ChromePIPEnergyAndPower.*

Change-Id: Idc352a76eba4c0161b451a8923f76a5ec5395454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2988856
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Avery Musbach <amusbach@chromium.org>