camera: hal3: Run a pre-init test for cros_camera_test
When a DUT has a hardware issue that does not let it run a test we do
not get an error message describbing the core issue, this is: error
initializing hardware, invalid number of cameras.

Before running the specific cros_camera_test we run a test on the core
functionality of the camera that will show an error message that can be
clusterised by crosmon.

BUG=b:191912349
TEST=rmmod uvcvideo; tast run $DUT camera.HAL3Recording

Change-Id: I28bdc31bf45ba344c18f6c452eb6a243e3c43ff9