Modify arc.OobeArc tast test to add the switch --force-launch-browser.
The Chrome browser is not automatially launched due to the full restore
requirement. This affects the arc.OobeArc tast test to show the OS
setting for the play store remove button. So add the switch
--force-launch-browser to force launching the chrome browser to show
the OS setting, so that the tast test can find the 'Remove Google Play
Store' button.

BUG=b:191210020
TEST=Ran tast test arc.OobeArc.vm

Change-Id: Ib63a3094b4696e5b7798ebd536b811c1ddc94add