wifi.Reset: Add timeout for suspend.
Suspend and reconnect share the same timeout interval,
so it's possible that suspend takes too much time and
leads to "failed to reconnect" error. By setting a
timeout for suspend, we get a clearer failure reason
telling whether it's a suspend issue or not.

BUG=b:175602523
TEST=`tast run ${DUT} wifi.Reset`
On board banon, 4 passed and 6 failed with
"failed to suspend: context deadline exceeded"

Change-Id: I8332d9ec4732a132a8c89d6ffc6ddf7a68304d15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2992734
Tested-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Will Donnelly <wgd@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>