tast-tests: Reduce flake of arc.PIPRoundedCornersUnderlay.vm
This CL is for reducing flake of arc.PIPRoundedCornersUnderlay.vm. That
test is intended to check that with the PipRoundedCorners feature, an
ARC PIP video is promoted to overlay using the underlay overlay strategy
or not at all. Referring to enum OverlayStrategies as defined in
tools/metrics/histograms/enums.xml in the chromium code base, the test
checks that on each frame, the overlay strategy is 1 (No overlay) or 4
(Underlay). It sometimes fails because there are some frames where the
overlay strategy is 6 (All overlays failed) or 7 (Overlays failed min).
However, there is not really anything wrong with applying those overlay
strategies to an ARC PIP video. So with this CL, the test will accept
them. For simplicity, arc.PIPRoundedCornersUnderlay will do the same.

BUG=b:184821569
TEST=tast run <DUT> arc.PIPRoundedCornersUnderlay*

Change-Id: I31230c72ca5f631ac10d89671ff13c90c69ab4df