upstart: Pass extra arguments in a structured way
StartJob and RestartJob accept extra arguments to pass to an upstart
job. They are currently represented as ...string, each of which is
formatted as "KEY=VALUE", an identical format as ones passed to
initctl(1). This approach lacks a proper abstraction over upstart CLI.

This patch modifies those functions to take ...Arg instead to be more
conforming to Go API standards. Callers can call WithArg to construct
Arg to be passed to functions, just like the following:

 upstart.StartJob(ctx, "foo", upstart.WithArg("key1", "value1"))

Note: I also considered API of the following style:

 func WithArgs(kvs map[string]string) []Arg

However, since it is rare to pass many arguments to upstart jobs,
I found this style gets more verbose, e.g.

 upstart.StartJob(ctx, "foo", upstart.WithArgs(map[string]string{"key1": "value1"})...)

BUG=b:192321435
TEST=fast_build.sh -T

Change-Id: I747ee09b1735982821fc046ec1bb8227f2b7e3b1