arc.RemovableMedia: Ensure the volume is mounted before accessed
arc.RemovableMedia.vm has been flaky because it often tries to access a
file inside a removable media volume before the volume is mounted on the
Android side.
This CL introduces a polling loop to wait for the volume to be mounted
in ARC before it is accessed.

BUG=b:192296595
TEST=tast run betty arc.RemovableMedia*

Change-Id: Ib0724117281f79ad070d8853f86f318275d0a8fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2995008
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>