tast-tests: syzcorpus: fix mkdtemp errors
Repros attempt to create temporary directories of the form
"./syzkaller.XXXXXX". When the repro is run over ssh, it attempts
to create the temporary directory in /root, and fails.

Fix this by changing the current directory to the temporary directory
the repro resides in before running it.

BUG=b:189825001
TEST=tast -verbose run localhost:2200 syzcorpus.Kvmtest

Change-Id: Ib9fbe5594bec27b9fc85681e8f4506a54b01f3f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2995488
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Commit-Queue: Zubin Mithra <zsm@chromium.org>