wifi.APSupportedRates: port from Tauto
This is mostly a faithful port of network_WiFi_APSupportedRates.

I currently omit the wlan.bssid filter used in the Tauto version (and
note a possible TODO instead), because that filter is a complex
wireshark rule which is a little hard to imitate directly in gopacket
(but we could probably imitate it by just directly filtering out ACKs
and BlockAcks if needed), and if our test AP behaved better, we wouldn't
need it at all. I'd consider running this in the lab for a while to see
how it behaves first.

BUG=b:160281713
TEST=`tast run ${HOST} wifi.APSupportedRates.*` on scarlet (skip) and
     others

Cq-Depend: chromium:2998304
Change-Id: Ie380508277ae55157a6e4c939c3cb439777d3d49