crostini: Increase retry limit for unsharing folders
This CL increases the retry limit for unsharing folders from 2 to 4.
This is intended to try and reduce flakiness of pita.FolderSharing.
As the retry interval is 5s, our final attempt will be 15s after the
first attempt (compared to 5s right now).

At the moment I don't think we have any better ways to wait for
seneschal to consider a folder as not in use than clicking the retry
button as we do.

BUG=chromium:192325276
TEST=Run pita.FolderSharing tast test

Change-Id: I7ca5bb2ddf231d04026d31a192297777b0a092d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2999352
Reviewed-by: Judy Wu <jinrongwu@google.com>
Tested-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>