tast-test: Wait for tinyproxy to write pid
tinyproxy can be configured to write its pid in a file. The pid is
required for setting up the network namespace of the proxy server.

Some tast reports show that we're trying to read the pid before it's
being written:
"2021-06-29T01:50:24.449594Z INFO tast[4580]:
    network.SystemProxyForSystemServices: Error at
    system_proxy_for_system_services.go:52: Failed to start a local
    proxy on the DUT: failed to setup the network namespace: failed to
    get proxy process pid: strconv.Atoi: parsing "": invalid syntax
2021-06-29T01:50:24.449925Z INFO tinyproxy[2085]: Initializing
    tinyproxy ..."

This CL implements retries to read the pid for at most 5 seconds.

BUG=none
TEST=tast run $DUT network.SystemProxyForSystemServices

Change-Id: I01d15fe20d89a3fa367c8bbea82594152f3e525e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2996186
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>