tast-tests: EnsureDUTBooted improved
If the DUT is already connected, avoid reconnecting, and just return.

There was a mistake that made EnsureDUTBooted return early with no err
if the servo is not working.

Fallback to a reset if power state != S0, or if the power state is not
available, or if waiting for boot times out.

BUG=b:191362994
TEST=test_that --board=${BOARD?} --model=${MODEL?} \
  ${DUT_HOSTNAME?} tast.firmware-smoke

Change-Id: I12a030c7c2eb465d0868ed9565bbc3be3c35a362
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2997806
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>