tast-tests: Port some a11y files to uiauto
This change is a part of the tast-wide effort to port existing
tests from chrome/ui to chrome/uiauto. We remove all uses of
chrome/ui within */a11y/a11y.go. Some changes also had to be made
to accessibility_event.go, so the file is still only partially
converted.

This change also adds a Matches() function to automation.go.
This method adds the ability to match a NodeInfo and a Finder,
similar to how a Node can be matched with a FindParams [1].

[1] https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/tast-tests/src/chromiumos/tast/local/chrome/ui/automation.go;l=501;drc=8379909dc26b1014dc9f3098f3f530a35f5cb638

BUG=b:190614781
TEST=tast -verbose run <IP> a11y.* arc.Accessibility*

Change-Id: Ic6533fca884cbfcdcfccd6119dce895721fee36b