CrOS AppList: Add tast tests for foldering actions in the launcher.
This CL adds tast tests to cover the following cases:
- Adding and removing apps from a folder.
- Filling a folder with apps to its maximum capacity.
- Deletion of folder by removing all apps.

Also, some utility functions have been added to launcher.go which are
useful for performing common actions in the launcher. These functions
include dragging icons on top of each other, dragging icons to the next
page, dragging icons out of a folder, and getting the size of a folder.

BUG=b:190654657
TEST=Ran all the app list foldering test

Change-Id: I64c7b6f11bac196b1d520f6ffa16ff1181a714a7