Adjust soak block timings to improve stress proportion
Driving the lifetime of the drive through stress testing is vital to test the endurance of the drives under test. The existing test only spent 25% of the overall duration running stress.
This change compresses the 'surfing' portion of the test and extends the 'stress' portion to be overall ~67% of the total test duration. This should more than double the overall wearout efficiency of the test.
Added a new "recovery" test block that exhibits a reduced load as well as 3 minutes of idle time.

BUG=b:192634438
TEST=tast -verbose run -var=tast_disk_size_gb=<disk_size_in_gb> <ip_address> storage.FullQualificationStress.stress

Change-Id: I5f2c726b2537129d5151e8a977e5f54cd1c5afce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3003023
Auto-Submit: David Jimenez <dvjimenez@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Alex Bergman <abergman@google.com>
Commit-Queue: David Jimenez <dvjimenez@chromium.org>
Tested-by: David Jimenez <dvjimenez@chromium.org>